// src/app/models/services-names.enum.ts
export enum ServicesNames {
  HARDWARE_MANAGEMENT = 'hardwareManagement',
  INDOOR_WEATHER_STATION = 'indoorWeatherStation',
  OUTDOOR_WEATHER_STATION = 'outdoorWeatherStation',
  PRINT_SERVER = 'printServer',
  MOTION_SENSORS = 'motionSensors',
  PARKING_CAMERAS = 'parkingCameras',
  MAIN_LIGHTS_CONTROL = 'mainLightsControl',
  DECORATIVE_LIGHTS_CONTROL = 'decorativeLightsControl',
}